+++ example/main.dart
@@ -5,16 +5,28 @@ main() async {
   var app = new Angel(reflector: MirrorsReflector());
 
   // Index route. Returns JSON.
-  app.get('/', () => 'Welcome to Angel!');
+  app.get('/', (req, res) => res.write('Welcome to Angel!'));
 
   // Accepts a URL like /greet/foo or /greet/bob.
-  app.get('/greet/:name', (String name) => 'Hello, $name!');
+  app.get(
+    '/greet/:name',
+    (req, res) {
+      var name = req.params['name'];
+      res.write('Hello, $name!');
+    },
+  );
 
   // Pattern matching - only call this handler if the query value of `name` equals 'emoji'.
-  app.get('/greet', (@Query('name', match: 'emoji') String name) => '😇🔥🔥🔥');
+  app.get(
+    '/greet',
+    ioc((@Query('name', match: 'emoji') String name) => '😇🔥🔥🔥'),
+  );
 
   // Handle any other query value of `name`.
-  app.get('/greet', (@Query('name') String name) => 'Hello, $name!');
+  app.get(
+    '/greet',
+    ioc((@Query('name') String name) => 'Hello, $name!'),
+  );
 
   // Simple fallback to throw a 404 on unknown paths.
   app.use((RequestContext req) async {