+++ test/parameter_meta_test.dart
@@ -1,7 +1,8 @@
 import 'dart:async';
-import 'dart:io';
-import 'package:angel_container/mirrors.dart';
 import 'dart:convert';
+import 'dart:io';
+
+import 'package:angel_container/mirrors.dart';
 import 'package:angel_framework/angel_framework.dart';
 import 'package:mock_request/mock_request.dart';
 import 'package:test/test.dart';
@@ -31,33 +32,33 @@ parameterMetaTests() {
     app = new Angel(reflector: MirrorsReflector());
     http = new AngelHttp(app);
 
-    app.get('/cookie', (@CookieValue('token') String jwt) {
+    app.get('/cookie', ioc((@CookieValue('token') String jwt) {
       return jwt;
-    });
+    }));
 
-    app.get('/header', (@Header('x-foo') String header) {
+    app.get('/header', ioc((@Header('x-foo') String header) {
       return header;
-    });
+    }));
 
-    app.get('/query', (@Query('q') String query) {
+    app.get('/query', ioc((@Query('q') String query) {
       return query;
-    });
+    }));
 
-    app.get('/session', (@Session('foo') String foo) {
+    app.get('/session', ioc((@Session('foo') String foo) {
       return foo;
-    });
+    }));
 
-    app.get('/match', (@Query('mode', match: 'pos') String mode) {
+    app.get('/match', ioc((@Query('mode', match: 'pos') String mode) {
       return 'YES $mode';
-    });
+    }));
 
-    app.get('/match', (@Query('mode', match: 'neg') String mode) {
+    app.get('/match', ioc((@Query('mode', match: 'neg') String mode) {
       return 'NO $mode';
-    });
+    }));
 
-    app.get('/match', (@Query('mode') String mode) {
+    app.get('/match', ioc((@Query('mode') String mode) {
       return 'DEFAULT $mode';
-    });
+    }));
 
     /*app.logger = new Logger('parameter_meta_test')
       ..onRecord.listen((rec) {