+++ CHANGELOG.md
@@ -40,3 +40,6 @@ take routes in the form of `FutureOr myFunc(RequestContext, ResponseContext res)
 * `@Middleware` now takes an `Iterable` of `RequestHandler`s.
 * `@Expose.path` now *must* be a `String`, not just any `Pattern`.
 * `@Expose.middleware` now takes `Iterable<RequestHandler>`, instead of just `List`.
+* `createDynamicHandler` was renamed to `ioc`, and is now used to run IoC-aware handlers in a
+type-safe manner.
+* `RequestContext.params` is now a `Map<String, dynamic>`, rather than just a `Map`.